dnsdist: Refactor QUIC tests so that they can be used for DoQ and DoH3
